fixed tests of PasswordRepository
